totalbill.cpp
Write a program that computes the tax and tip on a restaurant bill for a patron with a $44.50 meal charge.  The tax should be 6.75%of the meal cost.  The tip should be 15% of the total after adding the tax.  Display the meal cost, tax amount, tip amount, and total bill on the screen